Fix overloaded constructor selection under interop
* should not convert to optional parameters when argument count doesn't allow it
